+++ addons/purchase_stock/models/account_invoice.py
@@ -2,7 +2,8 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import fields, models
-from odoo.tools import float_compare
+from odoo.tools import float_compare, float_is_zero
+from odoo.tools.misc import groupby
 
 
 class AccountMove(models.Model):
@@ -144,7 +145,38 @@ class AccountMove(models.Model):
         if self._context.get('move_reverse_cancel'):
             return super()._post(soft)
         self.env['account.move.line'].create(self._stock_account_prepare_anglo_saxon_in_lines_vals())
-        return super()._post(soft)
+
+        # Create correction layer if invoice price is different
+        stock_valuation_layers = self.env['stock.valuation.layer'].sudo()
+        valued_lines = self.env['account.move.line'].sudo()
+        for invoice in self:
+            if invoice.sudo().stock_valuation_layer_ids:
+                continue
+            if invoice.move_type in ('in_invoice', 'in_refund', 'in_receipt'):
+                valued_lines |= invoice.invoice_line_ids.filtered(
+                    lambda l: l.product_id and l.product_id.cost_method != 'standard')
+        if valued_lines:
+            stock_valuation_layers |= valued_lines._create_in_invoice_svl()
+
+        for (product, company), dummy in groupby(stock_valuation_layers, key=lambda svl: (svl.product_id, svl.company_id)):
+            product = product.with_company(company.id)
+            if not float_is_zero(product.quantity_svl, precision_rounding=product.uom_id.rounding):
+                product.sudo().with_context(disable_auto_svl=True).write({'standard_price': product.value_svl / product.quantity_svl})
+
+        if stock_valuation_layers:
+            stock_valuation_layers._validate_accounting_entries()
+
+        posted = super()._post(soft)
+        # The invoice reference is set during the super call
+        for layer in stock_valuation_layers:
+            description = f"{layer.account_move_line_id.move_id.display_name} - {layer.product_id.display_name}"
+            layer.description = description
+            if layer.product_id.valuation != 'real_time':
+                continue
+            layer.account_move_id.ref = description
+            layer.account_move_id.line_ids.write({'name': description})
+
+        return posted
 
     def _stock_account_get_last_step_stock_moves(self):
         """ Overridden from stock_account.
